fix error

+++ database/migrations/2025_12_22_095844_add_approval_status_to_absensis_table.php
@@ -4,29 +4,29 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+return new class extends Migration {
 
-   
-public function up()
-{
-    if (!Schema::hasColumn('absensis', 'approval_status')) {
+    public function up()
+    {
+        if (!Schema::hasColumn('absensis', 'approval_status')) {
 
-        Schema::table('absensis', function (Blueprint $table) {
-            $table->enum('approval_status', [
-                'pending',
-                'disetujui',
-                'ditolak'
-            ])->default('pending');
-        });
+            Schema::table('absensis', function (Blueprint $table) {
+                $table->enum('approval_status', [
+                    'pending',
+                    'disetujui',
+                    'ditolak'
+                ])->default('pending');
+            });
 
+        }
     }
-}
 
-
-public function down(): void
-{
-    Schema::table('absensis', function (Blueprint $table) {
-        $table->dropColumn('approval_status');
-    });
-}
+    public function down()
+    {
+        if (Schema::hasColumn('absensis', 'approval_status')) {
+            Schema::table('absensis', function (Blueprint $table) {
+                $table->dropColumn('approval_status');
+            });
+        }
+    }
 };
